reverted removed imports of elephant's internal packages; download fim during the setup
